cancer-new-detection/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
whitenoise==6.4.0
tensorflow-cpu==2.12.0
numpy==1.23.5
